+++ backend/src/models/ObservationCollection.js
@@ -58,7 +58,6 @@ const ObservationCollectionSchema = new mongoose.Schema(
             type: String,
             required: true,
             unique: true,
-            immutable: true,
             select: false,
         },
         allowedOrigins: {